Use cursor into Chain instead of index
This is a speculative change and I'm just throwing the idea out there.
Whether or not this should be pursued should probably be determined by
benchmarks (not yet implemented) and anticipated use-cases.

The idea behind this is that the current approach of keeping an index
into a Chain probably isn't very efficient. Chain isn't optimized for
index-based usage. Since parsers generally want to pull an element at a
time and inspect it, keeping track of a "cursor" into the current "tail"
of the Chain may be more natural/efficient.

One potential downside of this is that when more data is supplied via
feed/prompt, we have to concatenate the data to both the `input` and the
`cursor`. Depending on the extent to which we have shared structure
(including after we start popping more elements off of the cursor), this
may result in larger space requirements (up to 2x).